### 2021-07-01 #3442 @whedon
[PRE REVIEW]: GeoClimate : a Geospatial processing toolbox for environmental and climate studies
**Submitting author:** @ebocher (<a href="http://orcid.org/0000-0002-4936-7079">Erwan Bocher</a>)
**Repository:** <a href="https://github.com/orbisgis/geoclimate/" target ="_blank">https://github.com/orbisgis/geoclimate/</a>
**Version:** 1.0.0-SNAPSHOT
