### 2021-07-06 orbisgis/geoclimate#604 @ebocher
JOSS article
A paper to present GeoClimate has been submitted to https://joss.theoj.org/

https://github.com/openjournals/joss-reviews/issues/3442
