### 2021-07-15 #3442 @ebocher
@kthyng
Just to inform you that the PDF compiles using github action : https://github.com/orbisgis/geoclimate/actions/runs/989591342
