### 2021-07-15 #3442 @arfon
@ebocher - please remove the blank ORCID lines from your paper by merging this pull request: https://github.com/orbisgis/geoclimate/pull/609

Also it looks like you're not citing your references properly. The section in your `paper.md` after the `# References` heading should be empty and citations to other works will automatically appear in there if you cite them in the paper text. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
