### 2021-07-15 e8308e1
Update according to openjournals/joss-reviews#3442
Remove all text from the reference section
Put all footnotes which were in the reference section within the article body
