+++ docs/articles/joss/paper.md
@@ -50,9 +50,9 @@ GeoClimate provides georeferenced morphological indicators as well as urban clas
 
 # Statement of need
 
-Urban spatial properties are useful to study the urban climate: (i) basic parameters such as building fraction or building height are needed as input of parametric urban climate models such as the Town Energy Balance [^teb] `[@masson2000]`, (ii) more sophisticated ones are clearly correlated to urban climate observations [^obs] and (iii) local climate classifications, useful for international comparisons, are mostly defined from urban spatial properties `[@stewart2012]`. Thus there is a need for tools dedicated to the calculation of urban spatial metrics.
+Urban spatial properties are useful to study the urban climate: (i) basic parameters such as building fraction or building height are needed as input of parametric urban climate models such as the Town Energy Balance teb^[https://github.com/teb-model/teb] `[@masson2000]`, (ii) more sophisticated ones are clearly correlated to urban climate observations^[Few examples: (i) the lower the Sky View Factor (SVF): the higher solar radiation are trapped by the urban canopy `[@bernabe2015]`, the higher the urban air temperature `[@lindberg2007]`, the lower the wind speed `[@johansson2016]` (ii) the higher the density of projected building facade in a given direction, the lower the wind speed within the urban canopy `[@hanna2002]`.] and (iii) local climate classifications, useful for international comparisons, are mostly defined from urban spatial properties `[@stewart2012]`. Thus there is a need for tools dedicated to the calculation of urban spatial metrics.
 
-In previous researches, scripts were developed to automatically calculate numerous indicators useful for urban climate applications `[@bocher2018]`. These scripts have been organized, improved and have been implemented within a Groovy library called GeoClimate. New urban properties and classifications algorithms have been added. GeoClimate also simplifies the access to geospatial data since it automatically downloads and organizes data from the world-wide OpenStreetMap database [^osm]. One of the current major limitations for the climate community to use this data is its lack of building height information `[@masson2020]`. Thus we have also added an algorithm to roughly estimate the height of each building missing this information.
+In previous researches, scripts were developed to automatically calculate numerous indicators useful for urban climate applications `[@bocher2018]`. These scripts have been organized, improved and have been implemented within a Groovy library called GeoClimate. New urban properties and classifications algorithms have been added. GeoClimate also simplifies the access to geospatial data since it automatically downloads and organizes data from the world-wide OpenStreetMap database^[https://www.openstreetmap.org]. One of the current major limitations for the climate community to use this data is its lack of building height information `[@masson2020]`. Thus we have also added an algorithm to roughly estimate the height of each building missing this information.
 
 This tool is first dedicated to urban climate researchers for modeling purpose: the output of GeoClimate can be directly used by urban climate models or by simple empirical models `[@bernard2017]`. It is also useful for any investigation dealing with urban climate issues (the calculation of the Local Climate Zone is for example of major interest as metadata for any urban climate study). The indicators calculated by GeoClimate can also be used for territory diagnostic and planning purpose for any spatial related question (climate, energy, biodiversity, pollution, socio-economy, etc.).
 
@@ -60,8 +60,8 @@ This tool is first dedicated to urban climate researchers for modeling purpose: 
 
 There is currently no software specifically designed for the calculation of geospatial indicators dedicated to urban climate. However, two softwares can currently be used to automatically perform some of the GeoClimate’s features:
 
[…]
