### 2021-07-15 #3442 @j3r3m1
> @ebocher - please remove the blank ORCID lines from your paper by merging this pull request: [orbisgis/geoclimate#609](https://github.com/orbisgis/geoclimate/pull/609)

Done

> Also it looks like you're not citing your references properly. The section in your `paper.md` after the `# References` heading should be empty and citations to other works will automatically appear in there if you cite them in the paper text. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)

The reason was that we put footnote in the reference section since it did not work otherwise. I removed them and add them in the body of the article but I am not sure it works as expected. You can let us know before we merge or we can merge and then see. The pull request is here:
orbisgis/geoclimate#610
