### 2021-07-15 b9bd338
Update paper.md
Modify according to @arfon  comment (references)
