### 2021-07-20 #3442 @elbeejay
Hi @ebocher, is there API documentation somewhere in your wiki? One of the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) is for the software to have some documentation of at least the core API functions. I apologize if I've missed seeing it in the wiki.

Secondly, can you recommend any reviewers that would be a good fit to review this work? A list of folks that have expressed a willingness to review for JOSS is available [here](https://bit.ly/joss-reviewers), but feel free to also recommend people not on this list. Please do not tag potential reviewers in this thread but rather provide their GitHub handle without the "@" symbol (e.g. "elbeejay" for me).

Thanks and feel free to reach out with any questions about the process.

- Jay
