### 2021-07-21 #3442 @ebocher
Hi @elbeejay ,

Thanks for the comments.
All GeoClimate algorithms are documented in the code using the javadoc syntax. Here you have an example : https://github.com/orbisgis/geoclimate/blob/master/geoindicators/src/main/groovy/org/orbisgis/geoclimate/geoindicators/BlockIndicators.groovy#L12

About potential reviewers there are  sjsrey, nuest, sgrieve, hugoledoux.

Note I will be OoO until middle of August

Best regards

Erwan
