### 2021-07-25 #3442 @omshinde
Hi @elbeejay! Thanks for the review invitation. I confirm that I do not hold any conflict of interest for this submission and would be happy to review this submission for JOSS. Regards!
