### 2021-07-26 #3541 @whedon
[REVIEW]: GeoClimate: a Geospatial processing toolbox for environmental and climate studies
**Submitting author:** @ebocher (<a href="http://orcid.org/0000-0002-4936-7079">Erwan Bocher</a>)
**Repository:** <a href="https://github.com/orbisgis/geoclimate/" target ="_blank">https://github.com/orbisgis/geoclimate/</a>
**Version:** 0.0.1
