### 2021-08-16 #3541 @elbeejay
:wave: Hello @abhishekvp @arbennett @omshinde we're at the halfway point for this review so I wanted to just check in with you all. Please feel free to reach out if you have any questions about the JOSS review process.
