+++ docs/README.md
@@ -2,7 +2,6 @@
 
 [![GitHub](https://img.shields.io/github/license/orbisgis/h2gis.svg)](https://github.com/orbisgis/geoclimate/blob/master/LICENSE.md) 
 [![Build Stat](https://img.shields.io/jenkins/s/http/jenkins.orbisgis.org/job/geoclimate.svg)](http://jenkins.orbisgis.org/job/geoclimate) 
-[![Codacy Badge](https://img.shields.io/codacy/grade/93899ea0675d43a2a3787ce5dd3c5595.svg)](https://www.codacy.com/app/orbisgis/geoclimate?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=orbisgis/geoclimate&amp;utm_campaign=Badge_Grade) 
 
  **A geospatial processing toolbox for environmental and climate studies** 
  
