### 2021-09-04 #3541 @omshinde
@elbeejay @ebocher

Review 1:
Thanks to the authors for their contributions. Please find my queries and suggestions below regarding the submission - "GeoClimate: a geospatial processing toolbox for environmental and climate studies".

Working Environment:
OS - Ubuntu 20.04.2 LTS

#### Functionality

Query: The download link as mentioned on [https://github.com/orbisgis/geoclimate/wiki/Download#geoclimate-for-command-line-interface](https://github.com/orbisgis/geoclimate/wiki/Download#geoclimate-for-command-line-interface) for the [Geoclimate.jar](http://jenkins.orbisgis.org/job/geoclimate-with-dependencies/lastSuccessfulBuild/artifact/geoclimate/target/geoclimate-1.0.0-SNAPSHOT-jar-with-dependencies.jar) redirects to file not found for me. Could you please recheck and confirm? I hope I am not missing anything here. Thanks!

I will continue my review for functionality and documentation after receiving response in this regard.

#### Software Paper
##### References
1. Minor Suggestion - The authors could consider citing the UMEP manual instead of putting the [link](https://umep-docs.readthedocs.io/en/latest/) as footnote in the Software paper.
    1.1 (Please refer: https://umep-docs.readthedocs.io/en/latest/Introduction.html#umep-how-to-cite)

    Lindberg F, Grimmond CSB, A Gabey, L Jarvi, CW Kent, N Krave, T Sun, N Wallenberg, HC Ward (2019) Urban Multi-scale Environmental Predictor (UMEP) Manual. https://umep-docs.readthedocs.io/ University of Reading UK, University of Gothenburg Sweden, SIMS China
