### 2021-09-04 #3541 @ebocher
@omshinde
Thank you for pointing out this error. The page has been updated and the link is now pointing to the releases page.
