### 2021-09-09 #3541 @elbeejay
:wave: @abhishekvp @arbennett @omshinde wanted to ask how the reviews are coming along and remind you all to report any review comments or suggestions as GitHub issues in the `GeoClimate` repository. When doing so just remember to tag the JOSS review issue (this thread) so that it is referenced here.

As we are beyond the 6-week target deadline for reviews, please provide new estimates for when you expect to complete your first evaluations of this submission so that @ebocher and co-authors can be aware and prepared.
