### 2021-09-11 #3541 @omshinde
Hi @elbeejay! I am done with my reviews based on the above checklist. Please find my comments below. After the minor suggestions are addressed, I am good to approve the submission for publication with JOSS based on the review checklist. Thanks!

#### Working Environment
OS: Ubuntu 20.04.2 LTS
Java version: `openjdk version "1.8.0_292"`

#### Submission
The authors presented a geospatial processing toolbox for environmental and climate studies named as GeoClimate. GeoClimate enables computation of climate and environmental paramaters such as Sky View Factor, urban classifications such as Local Climate Zones based on vector based inputs. The submission shows the examples using OpenStreetMaps and French BD Topo 2.2 version databases.

#### Functionality

The installation instructions works perfectly and are easy to follow. The [tutorials](https://github.com/orbisgis/geoclimate/wiki/Tutorials-for-Linux) cover the details and contains step-by-step instructions to execute the algorithms. I could run the examples for the default city of `Pont-de-Veyle`. In order to check the reproducibility and efficacy of GeoClimate, the instructions were followed for a city in India - Mumbai. Additionally, it was also tested using the `bounding box` as input for the OSM workflow. Please find the outputs (screenshots of visualization of generated GeoJSONs using QGIS) below for these cases to validate if they are correct -

- Default example for `Pont-de-Veyle, France`

![Screenshot from 2021-09-11 16-44-30](https://user-images.githubusercontent.com/21292545/132948406-7c018ee9-a04d-4e82-81ff-d343ea88826a.png)

- Example of a locality in `Mumbai, India`

![Screenshot from 2021-09-11 17-34-49](https://user-images.githubusercontent.com/21292545/132948425-190b0344-065d-4172-9a97-013904a87024.png)

But, while executing the `Urban Typology by Random Forest (UTRF)` classification, I faced some issues. The output log follows as [here](https://pastebin.com/Sc6azJQ8). (Note: I did trying solving this issue by following the hack mentioned in the [FAQs](https://github.com/orbisgis/geoclimate/wiki/Frequently-Asked-Questions))

The Documentation is more than sufficient to get started working with GeoClimate. Though the API methods documentation is missing but I feel that could be added in future contributions. The Software paper is very well drafted as well.

Some minor suggestions:
1. In the [Bounding box tutorial](https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D), the description for [Bounding box method with BDTOPO v2.2](https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D#bounding-box-method-with-bdtopo-v22) could be removed if it is not ready yet. There is only a heading with no further steps to follow.

2. The authors could consider citing the UMEP manual instead of putting the link as footnote in the Software paper.
1.1 (Please refer: https://umep-docs.readthedocs.io/en/latest/Introduction.html#umep-how-to-cite)

```Lindberg F, Grimmond CSB, A Gabey, L Jarvi, CW Kent, N Krave, T Sun, N Wallenberg, HC Ward (2019) Urban Multi-scale Environmental Predictor (UMEP) Manual. https://umep-docs.readthedocs.io/ University of Reading UK, University of Gothenburg Sweden, SIMS China```

My remarks:
GeoClimate provides a great utility in computing the environmental and climate related parameters based on several indicators. Thanks to the authors for their efforts and for providing a neat software package to the scientific community.

Kind regards,
Rajat
