### 2021-09-12 orbisgis/geoclimate#622 @arbennett
JOSS review: Paper
# Paper review of "GeoClimate: a Geospatial processing toolbox for environmental and climate studies" by Erwan Bocher et al.

## Summary comments
The summary opens with describing how human land use and land cover change modify climate, but when introducing GeoClimate the connection isn't entirely clear. Particularly I think an even broader "lead-in" of what GeoClimate is. From only reading the summary I understand that GeoClimate provides data, but is it's main purpose to gather and synthesize the data to usable formats, or is it more of a tool where a user brings their own data?

All of this is answered in paragraphs 2 and 3 of the statement of need section, but I think a couple of high level sentences in the summary would be nice.

## Processing steps comments
On point 1, I was confused by "represented by 2.5 geometries". Do these geometries represent specifc features like the walls and roof or something else? Why the half geometry (and what does that actually mean)?

I found the discussion of the block and RSU spatial units (lines 83-91) a bit confusing. What are each used for? Are both only considering 2D, and do both share the same minimum sizes for partitioning of water and vegetation?

## Minor comments
Line 27: Should `teb` be in parentheses?
Line 27: "(ii) more sophisticated ones" -> does "ones" refer to parameters?
Figure 5: I think it would be helpful to say "Sky View Factor (SVF)" rather than just SVF, since that's only mentioned in a footnote previously
