### 2021-09-12 orbisgis/geoclimate#623 @arbennett
JOSS review: broken links in docs
All of the links on the [example workflow ](https://github.com/orbisgis/geoclimate/blob/v1.0.0-RC1/docs/for_users/execution_examples/run_osm.md) are giving me 404 errors. For example : https://github.com/orbisgis/geoclimate/blob/master/processingchain/src/test/resources/org/orbisgis/orbisprocess/geoclimate/processingchain/config/osm_workflow_envelope_folderoutput.json
