### 2021-09-12 #3541 @arbennett
Hi @elbeejay - thanks for the reminder. My completed review is below.

@ebocher - thanks for all your (and your team's) efforts in developing GeoClimate. While a bit outside of my own area of expertise I was able to understand the goals and & reasons for developing GeoClimate quite easily. I was also able to run through some of the examples that you've provided in your documentation. Overall I think the paper was well written and the software functions well. I have only a couple of minor things that I would like to see changed. They are below:

- [ ] I provided a full review of the paper in this issue, along with some clarifications I'd like to see before accepting: https://github.com/orbisgis/geoclimate/issues/622
- [ ] Some of the links in the documentation were not working: https://github.com/orbisgis/geoclimate/issues/623
