### 2021-09-13 orbisgis/geoclimate#623 @ebocher
Thanks for the comment but this link is related to an old version. You must used now the wiki doc and the last release 0.0.1
