### 2021-09-13 orbisgis/geoclimate#622 @ebocher
Thanks for the comments we'll process them quickly
