### 2021-09-13 orbisgis/geoclimate#622 @ebocher
@j3r3m1 @ELSW56 @gpetit @franetibe @SPalominos
