### 2021-09-13 orbisgis/geoclimate#622 @ebocher
Thanks @omshinde  for the review.

https://github.com/openjournals/joss-reviews/issues/3541#issuecomment-917406606
