### 2021-09-13 #3541 @ebocher
@abhishekvp

>     * Some of the links in the documentation were not working:  JOSS review: broken links in docs orbisgis/geoclimate#623

The documentation you mention is ouf of dated. Can you tell me how you found it, from a link in the wiki doc? So I will be able to fix it.
