### 2021-09-14 df16304
Update to integrate JOSS review from @arbennett
This PR intregates some changes : 

1. Replace 2.5 geometries by 2D geometries. Wall and roof values are attributes to the spatial feature.
2. Clarify Block scale definition
2. Clarify RSU definition
