### 2021-09-14 #3541 @ebocher
> But, while executing the `Urban Typology by Random Forest (UTRF)` classification, I faced some issues. The output log follows as [here](https://pastebin.com/Sc6azJQ8). (Note: I did trying solving this issue by following the hack mentioned in the [FAQs](https://github.com/orbisgis/geoclimate/wiki/Frequently-Asked-Questions))

@omshinde
Thanks for the stack trace. It seems a pb with your network. Could you check if you have access to https://github.com/orbisgis/geoclimate/raw/master/models/URBAN_TYPOLOGY_BDTOPO_V2_RF_2_1.model
GeoClimate downloads this file to install it on your machine and run the machine learning algorithm.
