### 2021-09-14 #3541 @ebocher
@omshinde

>  Though the API methods documentation is missing but I feel that could be added in future contributions.
GeoClimate is a set of Groovy scripts so you are right there is no API but all scripts are fully documented. Here an example to apply the  machine learning algorithm :  https://github.com/orbisgis/geoclimate/blob/5bd90c2eeb36e626e5caf1b64bdae83c668e446a/geoindicators/src/main/groovy/org/orbisgis/geoclimate/geoindicators/TypologyClassification.groovy#L565
