### 2021-09-14 67fc430
Update paper.md
Add type of data currently considered in Geoclimate in the summary
