### 2021-09-14 49a7d06
Update paper.md according to @arbennett
Modification to the JOSS article according to minor comments in #622.
