### 2021-09-14 #3541 @ebocher
>     1. In the [Bounding box tutorial](https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D), the description for [Bounding box method with BDTOPO v2.2](https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D#bounding-box-method-with-bdtopo-v22) could be removed if it is not ready yet. There is only a heading with no further steps to follow.

The tutorial has been updated to integrate   BDTOPO v2.2 BBOX case.  see https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D
