### 2021-09-15 orbisgis/geoclimate#622 @ebocher
@arbennett , @omshinde
All comments have been fixed in the paper and a new version with the changes is available here https://github.com/orbisgis/geoclimate/suites/3773956828/artifacts/92897033
