### 2021-09-15 #3541 @abhishekvp
@ebocher @elbeejay Please find my review and comments below:

Thanks to the authors for their contribution to the Open Source Community and for their interest in publishing their work in the Journal of Open Source Software.

### Research Paper

The paper is very well organized. It discusses the GeoClimate processing toolbox that has been developed to compute complex morphological indicators and urban classifications. The need for such a tool in the urban studies research area has been justified. The authors also put forth a sound comparison of GeoClimate with UMEP and the LCZ Generator - the other prominent tools in this research area. The discussion on the coding implementation, configuration and the output is elaborate with the relevant details being covered. The authors list the relevant references that have been referred for this work. Overall the paper is excellently articulated and seems a great fit for the Journal of Open Source Software.

### Code Repository and Documentation

I was able to locally test the GeoClimate library on my system.

**Working Environment:** OpenJDK Runtime Environment (build 11.0.11+9-Ubuntu-0ubuntu2.20.04)

The tutorial documentation on the repository's wiki page is comprehensive and serves the purpose. I was able to successfully test the GeoClimate library for - (1) OSM with Locality value as a parameter and (2) OSM with Bounding Box. The indicators and classifications were successfully generated as desired in both cases.

I have a couple of minor suggestions which I would like the authors to address, before we go ahead with approving the paper for publication:

1. Although the content for the header - [Bounding box method with BDTOPO v2.2](https://github.com/orbisgis/geoclimate/wiki/Bounding-box-case-%5BL%5D#bounding-box-method-with-bdtopo-v22) has been added, as rightly pointed out by @omshinde in his review, the configuration snippet in the `description` field still says `Processing OSM data`. I believe it would need to be modified to `Processing BD Topo v2 data` for ensuring consistency.
2. I observed that the tutorial pages for - (1) BDTopo 2.2 and (2) Bounding box case for the Windows operating system, have not been created and simply point to the Wiki Home page of the repository. To ensure consistency in the documentation, the authors are encouraged to either (1) create these 2 pages - they may have contain a 'Work in Progress' message for now and the actual content can be populated when it is ready or (2) delete the pages from the navigation panel.

In conclusion, I would like to convey my sincere gratitude to the authors for their valuable contribution to the Open Source community. I am positive that GeoClimate would find great use in numerous applications for researchers working in the area of Urban Climate Studies.
