### 2021-09-20 #3541 @ebocher
Dear @abhishekvp
Thanks for the comments.
All tutorial pages have been updated and there is no a documentation the Windows OS.

@elbeejay
It seems we are now on the road. All comments, suggestions, corrections  have been done in the paper or in the documentation.
