### 2021-09-23 #3541 @omshinde
Hi @elbeejay, @ebocher
Thank you to the authors for addressing the comments patiently and promptly. I am very happy with the submission and feel elated to accept the "GeoClimate: a Geospatial processing toolbox for environmental and climate studies" for publication in JOSS. Congratulations to the authors for their efforts in making a great contribution to the scientific community.

@elbeejay Thanks for the opportunity to review this submission.
Kind regards,
Rajat
