### 2021-09-23 #3541 @arbennett
Hi @elbeejay & @ebocher - I am also happy to say that I am ready to accept this submission for publication in JOSS! Thanks again to @ebocher & the rest of the team for the quick turnaround, great work!
