### 2021-09-24 #3541 @elbeejay
Thanks again to our reviewers. @ebocher and co-authors I've opened a minor PR recommending some slight edits to the grammar of the paper, please review and decide if you would like to incorporate those.

Once that decision is made, there are just a few items that need to be taken care of:

- [ ] Make a tagged release of the software and report the version tag for the archived code here
- [ ] Archive the reviewed code on Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Check that the archived code (e.g., Zenodo) has the correct metadata. The title should match the paper title, the author list should be correct, and you can include authors' ORCID if you wish
- [ ] Provide the DOI of the archived code here

Thanks, please just reply with the above items once they are complete and then I will be able to send this on to the EiC for publication.
