### 2021-09-26 #3541 @abhishekvp
> Hi all, thanks for reviewing this work, and thank you to @ebocher and co-authors for making revisions as necessary.
>
> At this time I would like to confirm with @omshinde, @arbennett and @abhishekvp that the revised software and documentation in its current form is acceptable to you. At this time one outstanding review issue [here](https://github.com/orbisgis/geoclimate/issues/622) has yet to be closed, so @arbennett please take a look at the revisions made since that was opened and close the issue or make additional suggestions as necessary.
>
> Thanks again all for your contributions to this review process.

Hi All!

I too am happy to state that the revised software and documentation in its current form is acceptable to me.

@elbeejay I am grateful to you for the opportunity to review this work.
