### 2021-09-27 1e10c11
Update JOSS paper after copy editing
Minor grammar revisions
