### 2021-09-28 #3541 @elbeejay
> > Make a tagged release of the software and report the version tag for the archived code here
>
> This is done here https://github.com/orbisgis/geoclimate/releases/tag/0.0.1 We are working to compile also the doc with the release.

Can you make a tagged release of the current codebase? 27 commits have been made since 0.0.1 and I think at least 1 modified the source code.

> > Archive the reviewed code on Zenodo or a similar service (e.g., figshare, an institutional repository)
> > Check that the archived code (e.g., Zenodo) has the correct metadata. The title should match the paper title, the author list should be correct, and you can include authors' ORCID if you wish
>
> We plan to use https://www.softwareheritage.org but I don't know if it is able to generate a DOI.
>

@openjournals/joss-eics can https://www.softwareheritage.org be used for this purpose?
