### 2021-09-28 #3541 @arfon
> @openjournals/joss-eics can https://www.softwareheritage.org be used for this purpose?

You're of course welcome to deposit the software with https://www.softwareheritage.org but you will also need to make an archive with somewhere that will give you a DOI (Figshare, Zenodo, Dryad). I realize this is extra work for you all but our current document-production toolchain requires a DOI for the archive.
