### 2021-09-28 #3541 @ebocher
@elbeejay
> Can you make a tagged release of the current codebase? 27 commits have been made since 0.0.1 and I think at least 1 modified the source code.

The commits are related to the doc  not the code. I have compiled a new version of the doc with the release 0.0.1 that includes the commits. So I think it's not necessary to upgrade geoclimate to 0.0.2
