### 2021-09-28 #3541 @ebocher
@arfon
> I realize this is extra work for you all but our current document-production toolchain requires a DOI for the archive.

Don't worry about that. We will used zenodo. Thanks to offer  JOSS  the community.
