### 2021-09-29 #3541 @ebocher
@danielskatz

> @ebocher - please change the metadata on the zenodo deposit to match the title and authors of the paper, and let us know when this is done

Do you want the same name as in the paper  ?
