+++ docs/articles/joss/paper.md
@@ -44,21 +44,21 @@ bibliography: paper.bib
 
 # Summary
 
-Human activities induce changes on land use and land cover. These changes are most significant in urban areas where topographic features (e.g. building, road) affect the density of impervious surface areas and introduce a range of urban morphological patterns. Those characteristics impact the energy balance and modify the climate locally (e.g. inducing the so-called Urban Heat Island phenomenon). 
+Human activities induce changes on land use and land cover. These changes are most significant in urban areas where topographic features (e.g., building, road) affect the density of impervious surface areas and introduce a range of urban morphological patterns. Those characteristics impact the energy balance and modify the climate locally (e.g., inducing the so-called Urban Heat Island phenomenon). 
 
-Therefore, there is a need for georeferenced morphological indicators as well as urban classifications (such as Local Climate Zones) that can be directly used for planning or as inputs of climate models. GeoClimate is dedicated to this purpose: it converts raw geographical data (OpenStreetMap and French BDTopo) into indicators useful for climate applications (sky view factor, vegetation fraction, etc.). However, its application is not limited to the climate field. The indicators calculated in Geoclimate can also be used for other diagnostic or planning purposes: studying the territory fragmentation, the influence of the urban fabric on pollution (noise or air chemical transport), the energy consumption, etc. GeoClimate is available as a free and open source geospatial software. 
+Therefore, there is a need for georeferenced morphological indicators as well as urban classifications (such as Local Climate Zones) that can be directly used for planning or as inputs of climate models. GeoClimate is dedicated to this purpose: it converts raw geographical data (OpenStreetMap and French BDTopo) into indicators useful for climate applications (sky view factor, vegetation fraction, etc.) However, its application is not limited to the climate field. The indicators calculated in GeoClimate can also be used for other diagnostic or planning purposes: studying the territory fragmentation, the influence of the urban fabric on pollution (noise or air chemical transport), the energy consumption, etc. GeoClimate is available as free and open source geospatial software. 
 
 # Statement of need
 
-Urban spatial properties are useful to study the urban climate: (i) basic parameters such as building fraction or building height are needed as input of parametric urban climate models such as the Town Energy Balance (TEB)^[https://github.com/teb-model/teb] [@masson2000], (ii) more sophisticated parameters are clearly correlated to urban climate observations^[Few examples: (i) the lower the sky view factor: the higher solar radiation are trapped by the urban canopy [@bernabe2015], the higher the urban air temperature [@lindberg2007], the lower the wind speed [@johansson2016] (ii) the higher the density of projected building facade in a given direction, the lower the wind speed within the urban canopy [@hanna2002].] and (iii) local climate classifications, useful for international comparisons, are mostly defined from urban spatial properties [@stewart2012]. Thus there is a need for tools dedicated to the calculation of urban spatial metrics.
+Urban spatial properties are useful to study the urban climate: (i) basic parameters such as building fraction or building height are needed as input of parametric urban climate models such as the Town Energy Balance (TEB)^[https://github.com/teb-model/teb] [@masson2000], (ii) more sophisticated parameters are clearly correlated to urban climate observations^[A few examples: (i) the lower the sky view factor, the higher solar radiation trapped by the urban canopy [@bernabe2015], the higher the urban air temperature [@lindberg2007], the lower the wind speed [@johansson2016]; (ii) the higher the density of projected building facade in a given direction, the lower the wind speed within the urban canopy [@hanna2002].], and (iii) local climate classifications, useful for international comparisons, are mostly defined from urban spatial properties [@stewart2012]. Thus there is a need for tools dedicated to the calculation of urban spatial metrics.
 
[…]
