### 2021-09-29 #3541 @danielskatz
@ebocher - I've suggested some minor changes in the paper in https://github.com/orbisgis/geoclimate/pull/635 - please merge this or let me know what you disagree with
