### 2021-09-29 #3541 @danielskatz
I just noticed that @arbennett didn't check the box for
>Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

@arbennett - can you check this box, or let us know if there was a problem?
